requirements.txt
python_resize_image==1.1.18
Pillow==6.1.0
